backend/requirements.txt
Django==4.2.18
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
drf-spectacular==0.28.0
Pillow==12.1.1
APScheduler==3.10.4
python-dotenv==1.2.2
gunicorn==23.0.0
reportlab==4.4.0
bleach==6.2.0
